# http/swagger.yml
openapi: "3.0.0"
info:
  title: Influx API Service
  version: 0.1.0
servers:
  - url: /v2
paths:
...
  /flux:
    get:
      responses:
        '200':
          description: Links for flux language server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FluxLinks"
        default:
          description: Any response other than 200 is an internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /flux/suggestions:
    get:
      responses:
        '200':
          description: Suggestions for next functions in call chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FluxSuggestions"
        default:
          description: Any response other than 200 is an internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /authorizations:
    get:
      tags:
        - Authorizations
      summary: List all authorizations
      parameters:
        - in: query
          name: userID
          schema:
            type: string
          description: filter authorizations belonging to a user id
        - in: query
          name: user
          schema:
            type: string
          description: filter authorizations belonging to a user name
      responses:
        '200':
          description: A list of authorizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Authorizations"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Authorizations
      summary: Create an authorization
      requestBody:
        description: authorization to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Authorization"
      responses:
        '201':
          description: authorization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Authorization"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /buckets:
    get:
      tags:
        - Buckets
      summary: List all buckets
      responses:
        '200':
          description: a list of buckets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Buckets"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Buckets
      summary: Create a bucket
      requestBody:
        description: bucket to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Bucket"
      responses:
        '201':
          description: Bucket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bucket"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /orgs:
    get:
      tags:
        - Organizations
      summary: List all organizations
      responses:
        '200':
          description: A list of organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organizations"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Organizations
      summary: Create an organization
      requestBody:
        description: organization to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
      responses:
        '201':
          description: organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  '/orgs/{orgId}':
    get:
      tags:
        - Organizations
      summary: Retrieve an organization
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
          description: ID of organization to get
      responses:
        '200':
          description: organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags:
        - Organizations
      summary: Update an organization
      requestBody:
        description: organization update to apply
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
          description: ID of organization to get
      responses:
        '200':
          description: organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks:
    get:
      tags:
        - Tasks
      summary: List tasks.
      description: Lists tasks, limit 100
      parameters:
        - in: query
          name: after
          schema:
            type: string
          description: returns tasks after specified ID
        - in: query
          name: user
          schema:
            type: string
          description: filter tasks to a specific user id
        - in: query
          name: organization
          schema:
            type: string
          description: filter tasks to a specific organization id
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tasks"
                  links:
                    $ref: "#/components/schemas/Links"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Tasks
      summary: Create a new task
      requestBody:
        description: task to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /users:
    get:
      tags:
        - Users
      summary: List all users
      responses:
        '200':
          description: a list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Users"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Users
      summary: Create a user
      requestBody:
        description: user to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        '201':
          description: user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Authorization:
      properties:
        id:
          readOnly: true
          type: string
        token:
          readOnly: true
          type: string
        owner:
          $ref: "#/components/schemas/Owners"
      required: [userID]
    Authorizations:
      type: array
      items:
        $ref: "#/components/schemas/Authorization"
    Bucket:
      properties:
        id:
          readOnly: true
          type: string
        owners:
          $ref: "#/components/schemas/Owners"
        name:
          type: string
        retentionPeriod:
          type: integer
          format: int64
      required: [organizationId, name, retentionPeriod]
    Buckets:
      type: array
      items:
        $ref: "#/components/schemas/Bucket"
    Link:
      type: object
      readOnly: true
      description: URI of resource.
      properties:
        href:
          type: string
          format: url
      required: [href]
    Links:
      type: object
      properties:
        next:
          $ref: "#/components/schemas/Link"
        self:
          $ref: "#/components/schemas/Link"
        prev:
          $ref: "#/components/schemas/Link"
      required: [self]
    Organization:
      properties:
        id:
          readOnly: true
          type: string
        name:
          type: string
        owners:
          $ref: "#/components/schemas/Owners"
      required: [name]
    Organizations:
      type: array
      items:
        $ref: "#/components/schemas/Organization"
    Owners:
      properties:
        users:
          $ref: "#/components/schemas/Users"
        organizations:
          $ref: "#/components/schemas/Organizations"
    Run:
      properties:
        id:
          readOnly: true
          type: string
        taskId:
          readOnly: true
          type: string
        status:
          readOnly: true
          type: string
          enum: [
            "scheduled",
            "executing",
            "failed",
            "success"
          ]
        scheduledFor:
          description: Time used for run's "now" option, RFC3339.
          type: string
          format: date-time
        startedAt:
          readOnly: true
          description: Time run started executing, RFC3339Nano.
          type: string
          format: date-time
        finishedAt:
          readOnly: true
          description: Time run finished executing, RFC3339Nano.
          type: string
          format: date-time
        requestedAt:
          readOnly: true
          description: Time run was manually requested, RFC3339Nano.
          type: string
          format: date-time
        error:
          $ref: "#/components/schemas/Error"
        log:
          readOnly: true
          description: Link to the full logs for a run.
          type: string
    Task:
      properties:
        id:
          readOnly: true
          type: string
        organization:
          description: The ID of the organization that owns this Task.
          type: string
        name:
          description: A modifiable description of the task.
          type: string
        status:
          description: The current status of the task. When updated to 'disabled', cancels all queued jobs of this task.
          type: string
          enum: [
            "enabled",
            "disabled"
          ]
        owners:
          $ref: "#/components/schemas/Owners"
        flux:
          description: The Flux script to run for this task.
          type: string
        every:
          readOnly: true
          description: A simple task repetition schedule; parsed from Flux.
          type: string
        cron:
          readOnly: true
          description: A task repetition schedule in the form '* * * * * *'; parsed from Flux.
          type: string
        last:
          $ref: "#/components/schemas/Run"
      required: [name, organization, flux]
    Tasks:
      type: array
      items:
        $ref: "#/components/schemas/Task"
    User:
      properties:
        id:
          readOnly: true
          type: string
        name:
          type: string
      required: [name]
    Users:
      type: array
      items:
        $ref: "#/components/schemas/User"
    FluxSuggestions:
      type: object
      properties:
        funcs:
          type: object
          properties:
            name:
              type: string
            params:
              type: object
    FluxLinks:
      type: object
      properties:
        ast:
          type: string
        self:
          type: string
        suggestions:
          type: string
...
    Error:
      properties:
        code:
          readOnly: true
          type: integer
          format: int32
        message:
          readOnly: true
          type: string
      required: [code, message]
...